Allowing the base path used by the build_bootstrap.php script to be configurable as an argument
This is especially useful if your entire Symfony vendor directory is a sunlink. The hardcoded path in build_bootstrap.php means that the target directory is actually the parent directory of where vendor directory symlink target is, not the actual project. This causes the bootstrap files to be built in the wrong directory.