Add xdebug.max_nesting_level recommendation
The Xdebug extension includes infinite recursion protection, which throws a PHP fatal error if function nesting exceeds a configurable limit.

Unfortunately the default limit is so low that it is often exceeded in normal execution by symfony projects (especially in the dev environment). This triggers a panicked and futile search for a coding problem where none exists.

This commit adds a recommendation to increase the max_nesting_level limit above the default if Xdebug is loaded.