Simplifying further: both of these checks are ONLY valid if you have the intl extension
The problem with checking for class_exists() Collator or Locale, is that the Symfony
Intl component provides stubs for these. If you instead used get_class('Collator', false),
this would also probably work. But, there is a small risk that the Collator stub is already
loaded before this, and checking for the classes is more obtuse anyways.